+++ cpplox/memory.h
@@ -1,14 +1,14 @@
 ﻿#pragma once
 
-inline size_t grow_capacity(size_t capacity)
+inline int grow_capacity(int capacity)
 {
 	return (capacity < 8) ? 8 : capacity * 2;
 }
 
-void* reallocate(void* ptr, size_t oldSize, size_t newSize);
+void* reallocate(void* ptr, int oldSize, int newSize);
 
 template<typename T>
-T* allocate(size_t count)
+T* allocate(int count)
 {
 	return static_cast<T*>(reallocate(nullptr, 0, sizeof(T) * count));
 }
@@ -20,14 +20,14 @@ void free(T* ptr)
 }
 
 template<typename T>
-T* grow_array(T* ptr, size_t oldCount, size_t newCount)
+T* grow_array(T* ptr, int oldCount, int newCount)
 {
 	auto newPtr = reallocate(ptr, sizeof(T) * oldCount, sizeof(T) * newCount);
 	return static_cast<T*>(newPtr);
 }
 
 template<typename T>
-void free_array(T* ptr, size_t oldCount)
+void free_array(T* ptr, int oldCount)
 {
 	reallocate(ptr, sizeof(T) * oldCount, 0);
 }
